+++ examples/texture-arrays/src/main.rs
@@ -379,7 +379,7 @@ impl wgpu_example::framework::Example for Example {
                 resolve_target: None,
                 ops: wgpu::Operations {
                     load: wgpu::LoadOp::Clear(wgpu::Color::BLACK),
-                    store: true,
+                    store: wgpu::StoreOp::Store,
                 },
             })],
             depth_stencil_attachment: None,