+++ tests/tests/root.rs
@@ -21,6 +21,7 @@ mod mem_leaks;
 mod occlusion_query;
 mod partially_bounded_arrays;
 mod poll;
+mod query_set;
 mod queue_transfer;
 mod resource_descriptor_accessor;
 mod resource_error;